+++ packages/controls/js/libs/border-radius-control/index.js
@@ -131,7 +131,7 @@ export default function BorderRadiusControl({
 							});
 						}}
 						defaultValue={value.all || ''}
-						placeholder="-"
+						placeholder="0"
 						size="small"
 						data-test="border-radius-input-all"
 					/>
@@ -219,9 +219,8 @@ export default function BorderRadiusControl({
 							className={controlInnerClassNames(
 								'border-corner-top-left'
 							)}
-							noBorder={true}
 							defaultValue={value.topLeft || ''}
-							placeholder="-"
+							placeholder="0"
 							onChange={(newValue) => {
 								setValue({
 									...value,
@@ -244,9 +243,8 @@ export default function BorderRadiusControl({
 							className={controlInnerClassNames(
 								'border-corner-top-right'
 							)}
-							noBorder={true}
 							defaultValue={value.topRight || ''}
-							placeholder="-"
+							placeholder="0"
 							onChange={(newValue) => {
 								setValue({
 									...value,
@@ -269,9 +267,8 @@ export default function BorderRadiusControl({
 							className={controlInnerClassNames(
 								'border-corner-bottom-left'
 							)}
-							noBorder={true}
 							defaultValue={value.bottomLeft || ''}
-							placeholder="-"
+							placeholder="0"
 							onChange={(newValue) => {
 								setValue({
 									...value,
@@ -294,9 +291,8 @@ export default function BorderRadiusControl({
 							className={controlInnerClassNames(
 								'border-corner-bottom-right'
 							)}
-							noBorder={true}
 							defaultValue={value.bottomRight || ''}
-							placeholder="-"
+							placeholder="0"
 							onChange={(newValue) => {
 								setValue({
 									...value,